+++ src/runtime_src/xdp/profile/database/dynamic_info/device_db.h
@@ -112,7 +112,7 @@ namespace xdp {
     { aie_db.addAIETimerSample(timestamp1, timestamp2, values);  }
 
     inline
-    void addAIEDebugSample(uint64_t col, uint64_t row, uint64_t rel, uint64_t abslt, uint64_t val)
+    void addAIEDebugSample(uint64_t col, uint64_t row, uint64_t rel, uint64_t abslt, uint32_t val)
     { aie_db.addAIEDebugSample(col,row,rel,abslt,val);  }
 
     inline std::vector<counters::Sample> getAIESamples()
